constructor fixed - checkpoint
